Homework 4 for CSE240 (C)
Functions for a program which accepts 5 strings and 1 integer as input.
The program will then use the integer value to encrypt the input
strings, sort the encrypted strings by ASCII value, decrypt the strings,
and then print the strings. Can check the frequency of any letter,
remove vowels from the string, check if string is a palindrome and add a
letter to a spot. Done using pointer operations.